requirements.txt
fastapi
uvicorn
pydantic
# llama-cpp-python
faiss-cpu
PyMuPDF
pdfplumber
pandas
tqdm
tabulate
sentence-transformers
pytesseract